Arreglos FrmregistroUsuario

+++ App administracion Taquilla/admin/admin/frmregistroUsuario.cs
@@ -1,4 +1,7 @@
-﻿using System;
+﻿/*Autor: Melvin A. Batz Farfan
+  Carné: 0901-12-1049
+  */
+using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
@@ -23,12 +26,14 @@ namespace admin
 
         private void bguardar_Click(object sender, EventArgs e)
         {
+           //Query para realizar la insercion para registrar a un nuevo empleado
             string query = string.Format("INSERT INTO MAREGISTROPERSONAL (vnombre, vapellido, vtipo, vusuario, vcontrasenia) VALUES('{0}','{1}','{2}','{3}','{4}')", txtnombre.Text, txtapellido.Text, cmbtipousuario.Text, txtusuario.Text, txtpassword.Text); 
            
             try
             {
 
-                dllConexion.dllConexion.inserta(query);
+                dllConexion.dllConexion.inserta(query); //Utilizacion de una dll para la conexion con la BD
+                txtidempleado.ResetText();
                 txtnombre.ResetText();
                 txtapellido.ResetText();
                 cmbtipousuario.ResetText();
@@ -36,8 +41,8 @@ namespace admin
                 txtpassword.ResetText();
               
                 MessageBox.Show("Usuario registrado Exitosamente");
-                //Ingreso de Usuario en laa bitacora
-                claseEmp.Bita(claseEmp.MyUserName, 2, "MAREGISTROPERSONAL");
+
+                claseEmp.Bita(claseEmp.MyUserName, 2, "MAREGISTROPERSONAL");//Registrar la insercion en la Bitacora de Ingresos
             }
             catch
             {
@@ -58,8 +63,9 @@ namespace admin
                 String usuario = "";
                 String contrasenia = "";
 
-                string sql = "SELECT IIDREGISTRO,VNOMBRE as 'Nombre Empleado',VAPELLIDO,VTIPO,VUSUARIO,VCONTRASENIA FROM MAREGISTROPERSONAL WHERE IIDREGISTRO ='" + txtbidemp.Text + "' OR VNOMBRE = '" + txtbnombre.Text + "' OR VAPELLIDO= '" + txtbapellido.Text + "'OR VTIPO='" +cmbbtipo.Text + "'OR VUSUARIO='" + txtbusuario.Text + "'OR VCONTRASENIA='" + txtbpassword.Text + "'";
-                MySqlCommand cmd = new MySqlCommand(sql, dllConexion.dllConexion.Conexion());
+                // query para realizar la busqueda de un empleado, asi mismo asignandole un alias para que no se muestren los campos de la BD
+                string sql = "SELECT IIDREGISTRO,VNOMBRE as 'Nombre Empleado',VAPELLIDO,VTIPO,VUSUARIO,VCONTRASENIA FROM MAREGISTROPERSONAL WHERE IIDREGISTRO ='" + txtbidemp.Text + "' OR VNOMBRE = '" + txtbnombre.Text + "' OR VAPELLIDO= '" + txtbapellido.Text + "''OR VUSUARIO='" + txtbusuario.Text + "'";
+                MySqlCommand cmd = new MySqlCommand(sql, dllConexion.dllConexion.Conexion());// Utilizacion de la dllconexion para realiar conexion con la BD
 
                 MySqlDataReader reader = cmd.ExecuteReader();
 
@@ -72,7 +78,7 @@ namespace admin
                 buscarempleado.Columns.Add("Password", typeof(string));
                 
 
-                buscarempleado.Rows.Clear();                              //Limpia datos de la Tabla
+                buscarempleado.Rows.Clear();   //Limpia datos de la Tabla
 
                 if (reader.Read())
                 {
@@ -83,7 +89,7 @@ namespace admin
                     usuario = reader.GetString(4).ToString();
                     contrasenia= reader.GetString(5).ToString();
                     
-                    buscarempleado.Rows.Add(idempleado, nombre, apellido, tipo, usuario, contrasenia);
+                    buscarempleado.Rows.Add(idempleado, nombre, apellido, tipo, usuario, contrasenia);//Asignamos el nombre de los campos que tendra en DatagridView
 
                     MessageBox.Show("Empleado Encontrado");
 
@@ -94,7 +100,7 @@ namespace admin
 
 
                 }
-                grdbuscarempleado.DataSource = buscarempleado;
+                grdbuscarempleado.DataSource = buscarempleado;//Llenamos el DataGridView con los datos del empleado seleccionado
             }
             catch (Exception m)
             {
@@ -102,10 +108,12 @@ namespace admin
             }
         }
 
-        private void baceptarPelicula_Click(object sender, EventArgs e)
+        private void baceptarEmpleado_Click(object sender, EventArgs e)
         {
+            //Boton que nos perimitira realizar una eliminacion o edicion de campos, luego de haber realizado una busqueda en la BD
             frmregistroUsuario regempleado = new frmregistroUsuario();
             
+           //Seleccionamos todos los campos que se encuentran en el DataGridView para ser enviados a los campos de la busqueda de empleados
             regempleado.txtidempleado .Text = this.grdbuscarempleado.CurrentRow.Cells[0].Value.ToString();
             regempleado.txtnombre.Text = this.grdbuscarempleado.CurrentRow.Cells[1].Value.ToString();
             regempleado.txtapellido.Text = this.grdbuscarempleado.CurrentRow.Cells[2].Value.ToString();
@@ -139,10 +147,12 @@ namespace admin
         private void beditar_Click(object sender, EventArgs e)
         {
             frmregistroUsuario regempleado = new frmregistroUsuario();
+            //Query para realizar una modificacion sobre un empleado 
             string updateRegistro = ("UPDATE MAREGISTROPERSONAL set vnombre = '" + txtnombre.Text + "', vapellido = '" + txtapellido.Text + "', vtipo = '" + cmbtipousuario.Text + "', vusuario = '" + txtusuario.Text + "', vcontrasenia = '" + txtpassword.Text +"' WHERE iidregistro = '" + txtidempleado.Text + "'");
             dllConexion.dllConexion.inserta(updateRegistro);
             MessageBox.Show("El empleado " + txtnombre.Text + " se actualizo correctamente");
             
+            //Limpiamos todas las cajas de texto
             txtidempleado.ResetText();
             txtnombre.ResetText();
             txtapellido.ResetText();
@@ -150,7 +160,11 @@ namespace admin
             txtusuario.ResetText();
             txtpassword.ResetText();
 
-            
+            //Activamos y Desactivamos ciertos botones
+            regempleado.bguardar.Enabled = true;
+            regempleado.beditar.Enabled = false;
+            regempleado.beliminar.Enabled = false;
+            regempleado.brefrescar.Enabled = false;            
           
         }
 
@@ -161,24 +175,34 @@ namespace admin
 
         private void beliminar_Click(object sender, EventArgs e)
         {
-            frmregistroUsuario regempleado = new frmregistroUsuario();
-            String delete = "Delete from MAREGISTROPERSONAL WHERE IIDREGISTRO ='" + txtidempleado.Text + "'";
-            dllConexion.dllConexion.inserta(delete);
-            MessageBox.Show("El empleado " + txtnombre.Text + " se eliminó correctamente");
-            txtidempleado.ResetText();
-            txtnombre.ResetText(); 
-            txtapellido.ResetText();
-            cmbtipousuario.ResetText();
-            txtusuario.ResetText();
-            txtpassword.ResetText();
+            if (txtnombre.Text.Length == 0 && txtapellido.Text.Length == 0 && cmbtipousuario.Text.Length == 0 && txtusuario.Text.Length == 0 && txtpassword.Text.Length == 0)
+            {
 
-            regempleado.bguardar.Enabled = true;
-            regempleado.beditar.Enabled = false;
-            regempleado.beliminar.Enabled = false;
-            regempleado.button1.Enabled = false;
-                    
+                MessageBox.Show("Debe ingresar todos los datos solicitados");
+            }
+            else
+            {
+                frmregistroUsuario regempleado = new frmregistroUsuario();
+                //Realizamos una eliminacion por medio del codigo del empleado.
+                String delete = "Delete from MAREGISTROPERSONAL WHERE IIDREGISTRO ='" + txtidempleado.Text + "'";
+                dllConexion.dllConexion.inserta(delete);
+                MessageBox.Show("El empleado " + txtnombre.Text + " se eliminó correctamente");
+                
+                //Limpiamos todas las cajas de Texto
+                txtidempleado.ResetText();
+                txtnombre.ResetText();
+                txtapellido.ResetText();
+                cmbtipousuario.ResetText();
+                txtusuario.ResetText();
+                txtpassword.ResetText();
+
+                regempleado.bguardar.Enabled = true;
+                regempleado.beditar.Enabled = false;
+                regempleado.beliminar.Enabled = false;
+                regempleado.brefrescar.Enabled = false;
+
+            }
         }
-
         private void bguardar_Click_1(object sender, EventArgs e)
         {
 
@@ -189,14 +213,15 @@ namespace admin
             }
             else
             {
-
+                //Insercion de todos los datos de un nuevo empleado
                 string query = string.Format("INSERT INTO MAREGISTROPERSONAL (vnombre, vapellido, vtipo, vusuario, vcontrasenia) VALUES('{0}','{1}','{2}','{3}','{4}')", txtnombre.Text, txtapellido.Text, cmbtipousuario.Text, txtusuario.Text, txtpassword.Text);
 
                 try
                 {
 
-                    dllConexion.dllConexion.inserta(query);
+                    dllConexion.dllConexion.inserta(query);//utilizacion de la dllconexion
 
+                    //Limpieza de las cajas de Texto
                     txtnombre.ResetText();
                     txtapellido.ResetText();
                     cmbtipousuario.ResetText();
@@ -215,9 +240,30 @@ namespace admin
 
         private void button1_Click_1(object sender, EventArgs e)
         {
+          //Limpieza de las cajas de texto para un refresh del formulario
+            txtidempleado.Text = "";
+            txtnombre.Text = "";
+            txtapellido.Text = "";
+            txtusuario.Text = "";
+            txtpassword.Text = "";
+            cmbtipousuario.Text = "";
+            
             beditar.Enabled = false;
-            beliminar.Enabled = false;
+            beliminar.Enabled = true;
             bguardar.Enabled = true;
+        }
+
+        private void btnrefresh_Click(object sender, EventArgs e)
+        {
+            txtbidemp.Text = "";
+            txtbnombre.Text = "";
+            txtbapellido.Text = "";
+            txtbusuario.Text = "";
+
+
+            DataTable refrescar = (DataTable)grdbuscarempleado.DataSource;
+
+            refrescar.Clear(); //Limpiamos el Grid View cuando el usuario no desee realizar ninguna accion con los datos buscados.
         }
